+++ Localization/ruRU.lua
@@ -148,6 +148,7 @@ L.Settings.ClearOnLeave = "Очищать состав при выходе из 
 L.Settings.AutoRemoveRoles = "Автоматически удалять эти роли"
 L.Settings.SpotlightedHeader = "Выделенные"
 L.Settings.UnrosteredHeader = "Невыделенные"
+L.Settings.UnrosteredRoleFilter = "Показывать эти роли..."
 L.Settings.AddSpacer = "Добавить разделитель"
 L.Settings.NoSlots =
 "Пока никто не выделен. Добавьте участника группы из списка рядом или перетащите его сюда."